requirements.txt
gradio[mcp]>=5.45.0
python-dotenv>=1.0.0
Pillow>=9.0.0
requests>=2.28.0
dashscope>=1.24.5
# huggingface_hub>=0.20.0
#google-genai>=0.1.0